% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{setaria_viridis}
\alias{setaria_viridis}
\title{Size measurements for \emph{Setaria viridis}}
\format{
A data frame with 54 rows and 5 variables:
\describe{
\item{root_number}{A factor indicating which branches share the same root.}
\item{culm_length}{Numeric. Length of the culm above ground, measured in centimeters (cm).}
\item{panicle_width}{Numeric. Width of the panicle, measured in centimeters (cm).}
\item{panicle_length}{Numeric. Length of the panicle, measured in centimeters (cm).}
\item{leaf_number}{Integer. Number of leaves on the side branch.}
}
}
\usage{
setaria_viridis
}
\description{
This dataset contains measurements of various parts
and attributes for each branch of \emph{Setaria viridis}.
The variables are described as follows:
}
\keyword{datasets}
