% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{setaria_viridis}
\alias{setaria_viridis}
\title{Size measurements for Setaria viridis}
\format{
A data frame with 54 rows and 5 variables
\describe{
\item{root_number}{a factor denoting that it is growing from the same root}
\item{culm_length}{a number denoting culm length above ground (cm)}
\item{panicle_width}{a number denoting panicle width (cm)}
\item{panicle_length}{a number denoting panicle length (cm)}
\item{leaf_number}{an integer denoting number of leaves on the side branches (cm)}
}
}
\usage{
setaria_viridis
}
\description{
A dataset containing the length of each parts and
other attributes for each branch of Setaria viridis.
The variables are as follows:
}
\keyword{datasets}
