% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setariaviridis-package.R
\docType{package}
\name{setariaviridis-package}
\alias{setariaviridis}
\alias{setariaviridis-package}
\title{setariaviridis: Setaria Viridis Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Setaria viridis is a familiar weed.
    This package is data measured for each branch of Setaria viridis
    for practice data analysis. The data was collected
    from a wild Setaria viridis.
}
\author{
\strong{Maintainer}: Keisuke Ando \email{ando@maslab.aitech.ac.jp}

}
\keyword{internal}
