% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setariaviridis-package.R
\docType{package}
\name{setariaviridis-package}
\alias{setariaviridis}
\alias{setariaviridis-package}
\title{setariaviridis: Field-Collected Data of Green Foxtail (Setaria vilidis)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Setaria viridis (green foxtail) is a common weed. This package contains measurements from individual branches of a wild Setaria viridis plant collected near the author's home. The data is intended for use in data analysis practice.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NONONOexe/setariaviridis}
  \item \url{https://nononoexe.github.io/setariaviridis/}
  \item Report bugs at \url{https://github.com/NONONOexe/setariaviridis/issues}
}

}
\author{
\strong{Maintainer}: Keisuke ANDO \email{ando@maslab.aitech.ac.jp} (\href{https://orcid.org/0009-0002-4085-2067}{ORCID})

}
\keyword{internal}
